+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/F-PLN/MfdFmsFplnArr.tsx
@@ -252,6 +252,11 @@ export class MfdFmsFplnArr extends FmsPage<MfdFmsFplnArrProps> {
                 this.loadedFlightPlanIndex.get(),
                 isAltn,
               );
+              if (!isAltn) {
+                this.props.fmcService.master?.acInterface.deleteConstraintsAboveCruiseLevel(
+                  this.loadedFlightPlanIndex.get(),
+                );
+              }
             },
           },
         ];
@@ -266,6 +271,11 @@ export class MfdFmsFplnArr extends FmsPage<MfdFmsFplnArrProps> {
                 this.loadedFlightPlanIndex.get(),
                 isAltn,
               );
+              if (!isAltn) {
+                this.props.fmcService.master?.acInterface.deleteConstraintsAboveCruiseLevel(
+                  this.loadedFlightPlanIndex.get(),
+                );
+              }
             },
           };